:sparkles: Split the PO converter into two classes
The name was confusing and the class
had a lot of logic mixed.
